[COMPLETION] LCD Display Change on Switch Change
Now, I'm happy that I were able to do something that doesn't need a new function. Instead, I created a macro definition for us to clear the LCD screen when Switch Values is Changed or Updated. Layout is the only left for us to do!

Keep in mind that the placeholders are static values at the moment and we don't have any implementations that change the values.

We have a new bug occured! Please see issue regarding Blank Output....