:100: [COMPLETED IMPLEMENTATION] DS1302 (RTC Module) Added and Fully Functional~!
NOTE: Keep in mind that we just lost more than 300Bytes of SRAM because of this functioanlity. Rest assured that everything will be okay by trying to keep the SRAM more than > 5000 Bytes or else we are gonna face some consequence.

+ Added C-standard Library for Snprintf (stdio.h)
-+ Changed RTC Pins for Visual Seperation
+ Added RTC_MAX_BUFFER_SIZE Macro with Value of 50
+ Added String_BeautifyRTC Macro [NOTE: This will be intended for RTC only!, that's because converting strings for formality will make us lost a lot of bytes of SRAM for Arduino MEGA. It is usable or applicable but the consequence of losing more bytes is not acceptable. This will fix code readability but longer.]

- Removed CommsHeader_Status and Changed to RTC_Display_GetChipCurrentTime()

+Changed BatteryLevel to 100
+ Added RTC Initial Fully-Functional Functions... (Not gonna lie, I don't want to tell all the things that I added again because I'm tired.)

EXTRA: Some parts of RTC Functionality will be dependent on Serial Communication. So Serial Communication is basically one of the biggest part of this project. After fixing LCD Displays and finalizations, we are going right through it, before we try to go to MFRC522 and Optimizations. Also I learned how ... works for countless argument functions. :+1: