[CODE CLEANUP, 1 / ?] Defined Magic Numbers to Literal Significants
NOTE: This project is still flashable or runnable in any Arduino!
Keep in mind that changes done here due to messy code. There are no significant changes in the code, as of now. There will more other stages of clean up, this was just a sample at the moment. If I have time.