Clean Operation | Removed Stupid Overlayered Usage
What I meant is, I kept layering stupid namespaces and definitions where I just declared #pragma onces and it will all be fine.

I also forgot to mention, I removed a lot of namespaces on the recent commit due to annoying usage. Do you want to code or reference a function with a long name? Ofcourse not. So in the end I basically wasted most of my effort. But it benefits at some point. It helps me to divide variables and enums which is what I like.