Arduino Structure Change | De-Categorize Libraries
Why? Because, I'm receiving a lot of undefined error references from those libraries. And when I try to go down to the level of the directory of the library. I can now use them. THIS WAS CRITICAL the fact that I was initializing their classes and they can't recognize their .cpp source files. So in order to fix this, I have to remove THEM in the isolated folder and isolate them with .INO files. This kind of structure in ESP will be in effect once I was able to do something about it.