Development | Fixed Serial / System Doesn't Work
NOTE: We cannot push Serial.begin @ constructor of any class. I guess the reason why is because of the object being destroyed at the first place (in the end of the scope). If we remembered how python works in terms of OOP. We cannot access any references or objects created in that scope. Hence, unable to deploy the system which leads the system in the point of not able to output anything. This commit fixes it. SInce I fixed this one, I could push some of my works now :)